add tracy, propagate write/update errors from sqlite driver, better portable fixed-width integer handling

+++ include/savepoint/fwd.hpp
@@ -9,4 +9,4 @@ class SavepointEntity;
 class SavepointVersion;
 class SavepointVisitor;
 enum class SavepointDriver : uint8_t;
-enum class SavepointVisitorFlags : uint64_t;
+enum class SavepointVisitorFlags : uint32_t;
